+++ src/features/share-schedule/ui/index.ts
@@ -1,3 +1,3 @@
-export { default as ShareSchedule } from './share-calendar'
+export { default as ShareCalendar } from './share-calendar'
 export { default as ShareDetailedSchedule } from './share-detailed-schedule'
 export { default as ShareDetailedScheduleList } from './share-detailed-schedule-list'
